refactor(frontend): home types and schemas
